Day 12 of CF

+++ CalculatingFunction.java
@@ -2,7 +2,7 @@ import java.util.*;
 public class CalculatingFunction {
     public static void main(String[] args) {
         Scanner ob=new Scanner(System.in);
-        int n=ob.nextInt();
+        long n=ob.nextLong();
         if(n%2==0){
             System.out.println(n/2);
         }
